LineageRegistrationDialog: let it extend JFrame instead of JDialog
Microsoft Windows treats JDialog and JFrame differently. A JDialog does
not show up in the task manager. Extending JFrame therefore improves
usability.

There are also other reasons for this change. The LineageRegistrationDialog
does not behave like a dialog: It is not modal. There are no Ok or Cancel
buttons. There's only a close button. It's meant to stay open for a longer
duration, while interacting with other windows. It therefor behaves more
like a JFrame.