LineageRegistrationPlugin: change which cells are tagges as flipped.
Previously those cells that had differently sorted child cells where
tagged. This commit changes the code such that the child cells themselves
get tagged. Which is hopefully more intuitively to understand.

Incoming edges are tagged too.

The order of the buttons in the LineageRegistrationFrame is changed.
Tagging cells is usually the first action a user would want the use.
The buttons for tagging cells are therefor now in the top row.