LineageRegistration: use dividing spots for coordinate transform estimation
The algorithm previously used the first spot of each root cell to estimate
a coordinate transform between the two embryos. But we noticed for
example for Macrostomum that the cells move a lot before cell division
So which cell location (spot) along the trajectory of a cell would
be used for estimating the coordinate transform depends on the time at which
the recording starts. So it was basically random.

Now the algorithm uses the position of the dividing spot, which is better
defined.