Mastodon Git: don't initialize repository when sharing a project
It is very easy to initialize a repository when using github or gitlab.
And is impossible to go back to the uninitialized state. So it is better
for the "share project" command to assume that the github / gitlab repo
was already created. That way we also don't need to decide whether to
call the default branch "main" or "master".