Mastodon Git: clicking cancel on username, password dialog is handled better
Previously: The user clicking cancel on the username & password dialog
inderectly triggered a TransportException. The exception message would
be shown in an error dialog, and the stack trace was printed on the
console.

Now: The user clicking cancel throws a CancellationExcpetion. The
exception is catched at an appropriate location and no confusing
error dialog or stack trace is shown.