TomancakPlugins: fix merge datasets plugin, show result of merge operation
The plugin was recently changed such that it would no longer display the
result of the merge operation. The user was no longer to inspect or save
the result of the merge.

This is now fixed. The "Merge Datasets" plugin now closes the currently
open Mastodon windows and opens a new Mastodon with the results of the
merge operation.